Rename from 'utils' to 'nsgenerator'
