+++ testutil/testnode/full_node_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/celestiaorg/celestia-app/pkg/utils"
+	"github.com/celestiaorg/celestia-app/pkg/nsgenerator"
 	"github.com/stretchr/testify/suite"
 	tmrand "github.com/tendermint/tendermint/libs/rand"
 	coretypes "github.com/tendermint/tendermint/rpc/core/types"
@@ -75,7 +75,7 @@ func (s *IntegrationTestSuite) Test_Liveness() {
 
 func (s *IntegrationTestSuite) Test_PostData() {
 	require := s.Require()
-	_, err := s.cctx.PostData(s.accounts[0], utils.RandomValidNamespace(), tmrand.Bytes(100000))
+	_, err := s.cctx.PostData(s.accounts[0], nsgenerator.RandomMessageNamespace(), tmrand.Bytes(100000))
 	require.NoError(err)
 }
